DB1000N | Allow restarts only on Linux & Darwin systems
